Move components in the test directory
