+++ packages/formik/src/__tests__/components/index.ts
@@ -1,5 +1,4 @@
 import Button from './Button';
-import Field from './FieldAdapter';
 import Form from './Form';
 import Option from './Option';
 import RenderCheckBox from './RenderCheckBox';
@@ -7,4 +6,4 @@ import RenderField from './RenderField';
 import RenderRadio from './RenderRadio';
 import RenderSelect from './RenderSelect';
 
-export { Button, Field, Form, Option, RenderRadio, RenderCheckBox, RenderField, RenderSelect };
+export { Button, Form, Option, RenderRadio, RenderCheckBox, RenderField, RenderSelect };
